+++ packages/components/src/scroll-list/helpers/index.ts
@@ -1,1 +1,2 @@
 export * from './get-item-sizes';
+export * from './get-rendered-indexes';
